+++ CHANGELOG.md
@@ -28,6 +28,7 @@ CHANGELOG
  * Allow configuring Webhook's header names and signing algo
  * Deprecate registering console commands by overriding `Bundle::registerCommands()`, use the `#[AsCommand]` attribute or the `console.command` service tag instead
  * Deprecate calling `FrameworkExtension::load()` directly without first loading `ServicesBundle`'s extension
+ * Add support for the `#[RateLimit]` attribute on controllers when `framework.rate_limiter` is configured
 
 8.0
 ---